个人图片空间开发完成

+++ src/main/java/pres/yj/model/entity/Picture.java
@@ -120,7 +120,7 @@ public class Picture implements Serializable {
     /**
      * 空间 id
      */
-    private Integer spaceId;
+    private Long spaceId;
 
 
     @Serial
